% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimatepast_RKI_timeseries.R
\name{estimatepast_RKI_timeseries}
\alias{estimatepast_RKI_timeseries}
\title{estimatepast_RKI_timeseries}
\usage{
estimatepast_RKI_timeseries(
  x,
  ...,
  prop_death = 0.01,
  mean_days_until_death = 17,
  doubling_time = 4
)
}
\arguments{
\item{x}{data.frame. RKI data as downloaded with \code{\link{get_RKI_timeseries}}}

\item{...}{variable names. One or multiple grouping columns of x, so Bundesland, Landkreis, Geschlecht or Altersgruppe}

\item{prop_death}{numeric. Probability of death}

\item{mean_days_until_death}{integer. Mean number of days from infection to death (in case of death)}

\item{doubling_time}{numeric. Mean number of days for the number of infected to double}
}
\description{
estimatepast_RKI_timeseries
}
\examples{
rki_timeseries <- get_RKI_timeseries()

estimatepast_RKI_timeseries(rki_timeseries)
estimatepast_RKI_timeseries(rki_timeseries, Bundesland, Geschlecht)

}
