% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{ew_laender}
\alias{ew_laender}
\title{ew_laender}
\format{A data frame with 16 rows and 6 variables.
\itemize{
  \item \bold{Bundesland}
  \item \bold{FlaecheKm2}
  \item \bold{EwGesamt}
  \item \bold{EwMaennlich}
  \item \bold{EwWeiblich}
  \item \bold{EwProKm2}
}
For further information about the values see the documentation on https://www.destatis.de}
\description{
Source: https://www.destatis.de/DE/Themen/Laender-Regionen/Regionales/Gemeindeverzeichnis/Administrativ/04-kreise.html
Licence: Deutschland 2.0
}
\seealso{
Other context_data: 
\code{\link{ew_kreise}},
\code{\link{hospital_beds}}
}
\concept{context_data}
