% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_RKI_spatial.R
\name{get_RKI_spatial}
\alias{get_RKI_spatial}
\title{get_RKI_spatial}
\usage{
get_RKI_spatial(
  resolution = "Bundesland",
  urls = c(Bundesland =
    "https://opendata.arcgis.com/datasets/ef4b445a53c1406892257fe63129a8ea_0.zip",
    Landkreis =
    "https://opendata.arcgis.com/datasets/917fc37a709542548cc3be077a786c17_0.zip"),
  cache = T,
  cache_dir = tempdir(),
  cache_max_age = "today"
)
}
\arguments{
\item{resolution}{character. Resolution of the spatial data: "Bundesland" or "Landkreis"}

\item{urls}{named character vector. Data source urls}

\item{cache}{logical. Should the data be cached?}

\item{cache_dir}{character. Path to cache directory}

\item{cache_max_age}{numeric. Maximum age of cache in seconds}
}
\value{
An object of class sf with the dataset
}
\description{
Downloads the latest version of a COVID-19 spatial dataset by the Robert Koch Institute.
Please see the README for more information: \url{https://github.com/nevrome/covid19germany}
}
\examples{
\donttest{
rki_spatial_Bundesland <- get_RKI_spatial()
}

}
