% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_RKI_timeseries.R
\name{get_RKI_timeseries}
\alias{get_RKI_timeseries}
\title{get_RKI_timeseries}
\usage{
get_RKI_timeseries(
  cache = T,
  cache_dir = tempdir(),
  cache_max_age = 24 * 60 * 60
)
}
\arguments{
\item{cache}{logical. Should data be cached?}

\item{cache_dir}{character. Path to cache directory}

\item{cache_max_age}{numeric. Maximum age of cache in seconds}
}
\description{
get_RKI_timeseries
}
