% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_RKI_vaccination_progress.R
\name{get_RKI_vaccination_progress}
\alias{get_RKI_vaccination_progress}
\title{get_RKI_vaccination_progress}
\usage{
get_RKI_vaccination_progress(
  url = paste0("https://www.rki.de/",
    "DE/Content/InfAZ/N/Neuartiges_Coronavirus/Daten/Impfquotenmonitoring.xlsx",
    "?__blob=publicationFile"),
  raw_only = FALSE
)
}
\arguments{
\item{url}{character. Data source url}

\item{raw_only}{logical. Should covid19germany data cleaning not be applied}
}
\value{
A tibble with the dataset
}
\description{
Downloads the latest version of a COVID-19 vaccination progress dataset by the 
Robert Koch Institute.
Please see the README for more information: \url{https://github.com/nevrome/covid19germany}
}
\examples{
\donttest{
rki_vaccination_progress <- get_RKI_vaccination_progress()
}

}
