% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_RKI_timeseries.R
\name{group_RKI_timeseries}
\alias{group_RKI_timeseries}
\title{group_RKI_timeseries}
\usage{
group_RKI_timeseries(x, ...)
}
\arguments{
\item{x}{data.frame. RKI data as downloaded with \code{\link{get_RKI_timeseries}}}

\item{...}{character. "Bundesland", "Landkreis", "Geschlecht", "Altersgruppe"}
}
\description{
group_RKI_timeseries
}
