% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_factor.R
\name{growth_factor_RKI_timeseries}
\alias{growth_factor_RKI_timeseries}
\title{growth_factor_RKI_timeseries}
\usage{
growth_factor_RKI_timeseries(x, days = 1)
}
\arguments{
\item{days}{integer. Interval of days. Defaults to 1.}

\item{.data}{data.frame. Restructured RKI data}
}
\description{
Calculate the growth factors (Vervielfaeltigung pro Tag) in RKI
data which have been restructered with group_RKI_timeseries(...)
before. The function works on any grouping made with
group_RKI_timeseries().
}
\details{
The growth factor is defined as the
}
\examples{
timeseries <- get_RKI_timeseries()

republik <- group_RKI_timeseries(timeseries)

wachstum_republik <- growth_factor_RKI_timeseries(republik)

mutate(wachstum_republik, Meldedatum = as.Date(Meldedatum)) \%>\%
    ggplot() +
    geom_line(aes(x = Meldedatum, y = Wachstumsfaktor)) +
    geom_smooth(aes(x = Meldedatum, y = Wachstumsfaktor)) +
    scale_x_date(date_breaks = 'week', date_labels = '\%b \%d\n\%a')


bundeslaender <- group_RKI_timeseries(timeseries, Bundesland)

wachstum_bundeslaender <- growth_factor_RKI_timeseries(bundeslaender)

}
