% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{hospital_beds}
\alias{hospital_beds}
\title{hospital_beds}
\format{
A data frame with 16 rows and 7 variables.
\itemize{
  \item \bold{Bundesland}
  \item \bold{NumberHospital}
  \item \bold{NumberHospitalwithICU}
  \item \bold{NumberICUBed}
  \item \bold{NumberDaysICUBedinUse}
  \item \bold{NumberICUCase}
  \item \bold{NumberICUCasewithRespirator}
}
}
\description{
Source: \url{http://www.gbe-bund.de/gbe10/f?f=328::Intensivstation}, 
Please see the README for more information: \url{https://github.com/nevrome/covid19germany}
}
\seealso{
Other context_data: 
\code{\link{ew_alter}},
\code{\link{ew_kreise}},
\code{\link{ew_laender}}
}
\concept{context_data}
