% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{ew_alter}
\alias{ew_alter}
\title{ew_alter}
\format{
A data frame with 6 rows and 4 variables.
\itemize{
  \item \bold{Altersgruppe}
  \item \bold{Landkreis}
  \item \bold{EwGesamt}
  \item \bold{EwMaennlich}
  \item \bold{EwWeiblich}
}
}
\description{
Source: \url{https://www-genesis.destatis.de/genesis/online}, 
Licence: Deutschland 2.0, 
Please see the README for more information: \url{https://github.com/nevrome/covid19germany}
}
\seealso{
Other context_data: 
\code{\link{ew_kreise}},
\code{\link{ew_laender}},
\code{\link{hospital_beds}}
}
\concept{context_data}
