% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{ew_kreise}
\alias{ew_kreise}
\title{ew_kreise}
\format{
A data frame with 401 rows and 8 variables.
\itemize{
  \item \bold{IdLandkreis}
  \item \bold{Landkreis}
  \item \bold{NUTS3}
  \item \bold{FlaecheKm2}
  \item \bold{EwGesamt}
  \item \bold{EwMaennlich}
  \item \bold{EwWeiblich}
  \item \bold{EwProKm2}
}
}
\description{
Source: \url{https://www.statistikportal.de/de/bevoelkerung/flaeche-und-bevoelkerung}, 
Licence: Deutschland 2.0, 
Please see the README for more information: \url{https://github.com/nevrome/covid19germany}
}
\seealso{
Other context_data: 
\code{\link{ew_alter}},
\code{\link{ew_laender}},
\code{\link{hospital_beds}}
}
\concept{context_data}
