% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{ew_laender}
\alias{ew_laender}
\title{ew_laender}
\format{
A data frame with 16 rows and 6 variables.
\itemize{
  \item \bold{Bundesland}
  \item \bold{AreaKm2}
  \item \bold{PopulationTotal}
  \item \bold{PopulationMale}
  \item \bold{PopulationFemale}
  \item \bold{PopulationperKm2}
}
}
\description{
Source: \url{https://www.destatis.de/DE/Themen/Laender-Regionen/Regionales/Gemeindeverzeichnis/Administrativ/04-kreise.html}, 
Licence: Deutschland 2.0, 
Please see the README for more information: \url{https://github.com/nevrome/covid19germany}
}
\seealso{
Other context_data: 
\code{\link{ew_alter}},
\code{\link{ew_kreise}},
\code{\link{hospital_beds}}
}
\concept{context_data}
