% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_RKI_timeseries.R
\name{group_RKI_timeseries}
\alias{group_RKI_timeseries}
\title{group_RKI_timeseries}
\usage{
group_RKI_timeseries(x, ...)
}
\arguments{
\item{x}{data.frame. RKI data as downloaded with \code{\link{get_RKI_timeseries}}}

\item{...}{variable names. One or multiple grouping columns of x, so Bundesland, Landkreis, Gender or Age}
}
\description{
Restructures the RKI data to a grouped time series data.frame. Days with no observations are added to make them 
explicit. Grouping can be done with one or multiple variables (see parameter \code{...}).
Please see the README for more information: \url{https://github.com/nevrome/covid19germany}
}
\examples{
\donttest{
rki_timeseries <- get_RKI_timeseries()

group_RKI_timeseries(rki_timeseries, Bundesland, Gender)
}

}
