% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_RKI_timeseries.R
\name{plot_RKI_timeseries}
\alias{plot_RKI_timeseries}
\title{plot_RKI_timeseries}
\usage{
plot_RKI_timeseries(
  x,
  group = "Bundesland",
  type = "KumAnzahlFall",
  label = T,
  logy = F
)
}
\arguments{
\item{x}{data.frame. RKI data as downloaded with \code{\link{get_RKI_timeseries}}}

\item{group}{character. "Bundesland", "Landkreis", "Geschlecht", "Altersgruppe"}

\item{type}{character. Counter type}

\item{label}{logical}

\item{logy}{logical}
}
\description{
plot_RKI_timeseries
}
